requirements.txt
flask==2.3.3
flask-cors==4.0.0
groq>=0.30.0
python-dotenv==1.0.0
requests==2.31.0
streamlit==1.28.1
httpx>=0.24.0 